Suppress warnings

+++ include/noct/noct.h
@@ -96,8 +96,8 @@ struct rt_value {
 
 	/* Offset 8; */
 	union {
-		int i;
-		long l;
+		int32_t i;
+		int64_t l;
 		float f;
 		double lf;
 		struct rt_string *str;
@@ -286,7 +286,7 @@ bool
 noct_make_int(
 	NoctEnv *env,
 	NoctValue *val,
-	int i);
+	int32_t i);
 
 /*
  * Makes a long value.
@@ -296,7 +296,7 @@ bool
 noct_make_long(
 	NoctEnv *env,
 	NoctValue *val,
-	long l);
+	int64_t l);
 
 /*
  * Makes an integer value.
@@ -402,7 +402,7 @@ bool
 noct_get_int(
 	NoctEnv *env,
 	NoctValue *val,
-	int *i);
+	int32_t *i);
 
 /*
  * Retrieves an integer from a value with type checking.
@@ -414,7 +414,7 @@ bool
 noct_get_long(
 	NoctEnv *env,
 	NoctValue *val,
-	long *l);
+	int64_t *l);
 
 /*
  * Retrieves an integer from a value with type checking.
@@ -462,7 +462,7 @@ bool
 noct_get_string_len(
 	NoctEnv *env,
 	NoctValue *val,
-	uint32_t *len);
+	size_t *len);
 
 /*
  * Retrieves the string from a value with type checking.
@@ -944,7 +944,7 @@ noct_get_array_elem_check_int(
 	NoctValue *array,
 	uint32_t index,
 	NoctValue *val,
-	int *i);
+	int32_t *i);
 
 /*
  * Convenience function to retrieve a long element from an array
@@ -970,7 +970,7 @@ noct_get_array_elem_check_long(
 	NoctValue *array,
 	uint32_t index,
 	NoctValue *val,
-	long *l);
+	int64_t *l);
 
 /*
  * Convenience function to retrieve a float element from an array with
@@ -1130,7 +1130,7 @@ noct_set_array_elem_make_int(
 	NoctValue *array,
 	uint32_t index,
 	NoctValue *val,
-	int i);
+	int32_t i);
 
 /*
  * Convenience function to set a long element in an array.
@@ -1156,7 +1156,7 @@ noct_set_array_elem_make_long(
 	NoctValue *array,
 	uint32_t index,
 	NoctValue *val,
-	long l);
+	int64_t l);
 
 /*
  * Convenience function to set a float element in an array.
@@ -1258,7 +1258,7 @@ noct_get_dict_elem_check_int(
 	NoctValue *dict,
 	const char *key,
 	NoctValue *val,
-	int *i);
+	int32_t *i);
 
 /*
  * Convenience function to retrieve a long value associated with a
@@ -1285,7 +1285,7 @@ noct_get_dict_elem_check_long(
 	NoctValue *dict,
 	const char *key,
 	NoctValue *val,
-	long *l);
+	int64_t *l);
 
 /*
  * Convenience function to retrieve a float value associated with a
@@ -1448,7 +1448,7 @@ noct_set_dict_elem_make_int(
 	NoctValue *dict,
 	const char *key,
 	NoctValue *val,
-	int i);
+	int32_t i);
 
 /*
  * Convenience function to set a long value for a key in a dictionary.
@@ -1471,7 +1471,7 @@ noct_set_dict_elem_make_long(
 	NoctValue *dict,
 	const char *key,
 	NoctValue *val,
-	long l);
+	int64_t l);
 
 /*
  * Convenience function to set a float value for a key in a dictionary.
@@ -1560,7 +1560,7 @@ noct_get_arg_check_int(
 	NoctEnv *env,
 	uint32_t index,
 	NoctValue *val,
-	int *i);
+	int32_t *i);
 
 /*
  * Convenience function to retrieve a long function argument with
@@ -1583,7 +1583,7 @@ noct_get_arg_check_long(
 	NoctEnv *env,
 	uint32_t index,
 	NoctValue *val,
-	long *l);
+	int64_t *l);
 
 /*
  * Convenience function to retrieve a size_t function argument with
@@ -1734,7 +1734,7 @@ bool
 noct_set_return_make_int(
 	NoctEnv *env,
 	NoctValue *val,
-	int i);
+	int32_t i);
 
 /*
  * Convenience function to set a long return value for the current
@@ -1756,7 +1756,7 @@ bool
 noct_set_return_make_long(
 	NoctEnv *env,
 	NoctValue *val,
-	long l);
+	int64_t l);
 
 /*
  * Convenience function to set an integer return value for the current
